MindMapsApp/requirements.txt
requests
Flask
openai
rdflib
pydantic
python-dotenv
PyPDF2
python-docx